+++ lib/variables.ml
@@ -1739,6 +1739,9 @@ let vars_of_stroke_dashoffset (value : Properties.stroke_dashoffset) =
 let vars_of_stroke_dasharray (value : Properties.stroke_dasharray) =
   match value with Var v -> [ V v ] | _ -> []
 
+let vars_of_paint_order (value : Properties.paint_order) =
+  match value with Var v -> [ V v ] | _ -> []
+
 let vars_of_css_wide (value : Properties.css_wide) =
   match value with Var v -> [ V v ] | _ -> []
 
@@ -2183,6 +2186,7 @@ let vars_of_property : type a. a property -> a -> any_var list =
   | Stroke_miterlimit, value -> vars_of_stroke_miterlimit value
   | Stroke_dashoffset, value -> vars_of_stroke_dashoffset value
   | Stroke_dasharray, value -> vars_of_stroke_dasharray value
+  | Paint_order, value -> vars_of_paint_order value
   | Display, value -> vars_of_display value
   | Fill, value -> vars_of_svg_paint value
   | Flex, value -> vars_of_flex value